Add boss door and related scripts to level 3
Introduces 'puerta_boss' scene and script, as well as a trigger script for the boss door. Updates 'enemigo2_0' to include a pig grunt sound, and modifies 'escenario3' and 'tercer_nivel' scenes to integrate the new boss door and related elements. Also adds scripts for level 3's canvas layer and a new enemy type.